# Cache trimmed keys for @ and $