``` ini

BenchmarkDotNet=v0.13.5, OS=Windows 11 (10.0.22621.1265/22H2/2022Update/SunValley2)
AMD Ryzen 7 6800H with Radeon Graphics, 1 CPU, 16 logical and 8 physical cores
.NET SDK=7.0.201
  [Host]     : .NET 5.0.17 (5.0.1722.21314), X64 RyuJIT AVX2
  DefaultJob : .NET 5.0.17 (5.0.1722.21314), X64 RyuJIT AVX2
```

before 

|               Method |     Mean |     Error |    StdDev |   Gen0 | Allocated |
|--------------------- |---------:|----------:|----------:|-------:|----------:|
|       LogInformation | 3.141 μs | 0.0555 μs | 0.0519 μs | 0.1717 |   1.42 KB |
| LogInformationScoped | 4.421 μs | 0.0408 μs | 0.0381 μs | 0.2899 |   2.38 KB |

after

|               Method |     Mean |     Error |    StdDev |   Gen0 | Allocated |
|--------------------- |---------:|----------:|----------:|-------:|----------:|
|       LogInformation | 3.066 μs | 0.0348 μs | 0.0325 μs | 0.1640 |   1.36 KB |
| LogInformationScoped | 3.948 μs | 0.0320 μs | 0.0300 μs | 0.2823 |   2.32 KB |