@sungam3r 2023-03-09
I'd use cache for both cases. Ok with limiting size.